created a Context dir and a FormContext.js. Tested to see State from SiteBuildB is being held in ContextState
