+++ frontend/src/App.jsx
@@ -19,24 +19,27 @@ import ReviewAndSubmit from "./Pages/WhiteLabeling/ReviewAndSubmit";
 import OptionalQuestions from "./Pages/OptionalQuestionsPage/OptionalQuestions";
 import CriticalQuestions from "./Pages/CriticalQuestionsPage/CriticalQuestions";
 import OptionalContactInfo from "./Pages/OptionalContactPage/OptionalContact";
+import { FormProvider } from "./Context/FormContext";
 
 function App() {
   return (
     <div className="App">
       <Nav />
-      <Routes>
-        <Route exact path="/" element={<Home />} />
-        {/* <Route path="/DesignStart" element={<DesignStart />} /> */}
-        <Route path="/PartnerWhiteLabel" element={<PartnerWhiteLabel />} />
-        {/* <Route path="/SiteBuildA" element={<SiteBuildA />} /> */}
-        <Route path="/RequestSite" element={<RequestSite />} />
-        {/* <Route path="/SiteBuildC" element={<SiteBuildC />} /> */}
-        <Route path="/CriticalQuestions" element={<CriticalQuestions />} />
-        <Route path="/OptionalQuestions" element={<OptionalQuestions />} />
-        <Route path="/OptionalContactInfo" element={<OptionalContactInfo />} />
-        <Route path="/ReviewAndSubmit" element={<ReviewAndSubmit />} />
-        <Route path="/RequestConfirmation" element={<RequestConfirmation />} />
-      </Routes>
+      <FormProvider>
+        <Routes>
+            <Route exact path="/" element={<Home />} />
+            {/* <Route path="/DesignStart" element={<DesignStart />} /> */}
+            <Route path="/PartnerWhiteLabel" element={<PartnerWhiteLabel />} />
+            {/* <Route path="/SiteBuildA" element={<SiteBuildA />} /> */}
+            <Route path="/RequestSite" element={<RequestSite />} />
+            {/* <Route path="/SiteBuildC" element={<SiteBuildC />} /> */}
+            <Route path="/CriticalQuestions" element={<CriticalQuestions />} />
+            <Route path="/OptionalQuestions" element={<OptionalQuestions />} />
+            <Route path="/OptionalContactInfo" element={<OptionalContactInfo />} />
+            <Route path="/ReviewAndSubmit" element={<ReviewAndSubmit />} />
+            <Route path="/RequestConfirmation" element={<RequestConfirmation />} />
+        </Routes>
+      </FormProvider>
       <Footer />
     </div>
   );
