+++ frontend/src/Components/RequestSite/SiteBuildA/SiteBuildA.jsx
@@ -1,0 +1,8 @@
+import "./SiteBuildA.css";
+
+const SiteBuildA = (props) => {
+  return <>
+  </>;
+};
+
+export default SiteBuildA;
